Feat: created list of cats that show image
This is only partially completed.  I've added a list that you can
click on individual names to bring up the image of the cat and each
one can be clicked on, incrementing a counter on the image.

There is an issue though in that once you click on one of the cats
names, its number of clicks get incremented even when you are clicking
on a different cat.

This will need to be fixed in order to satisfy the requirements of the
project