Fix: overhauled the code to create and remove image divs
The previous problem was undoubtedly due to multiple image divs
occupying the same space, and each having their own click event handlers
which were created whenever they were displayed, and thus creating multiple
handlers for the same click.  Even setting display to none did not solve
this issue, and so in order to get it working I had to actually create and
remove the image div nodes each time a new list item was selected, so that
the event handlers wouldn't combine.  I was able to store the counts in a
separate array that was keyed based on the name of the cat.