Feat: updated code to Cat Clicker Premium Pro
This is a re-architecting of the app into an MVO architecture.
It is currently incomplete as it has some issues with the admin
section - still need to show/hid the admin form and the form fields
read the current cat object correctly but submitting only changes
the values temporarily and then reloads everything with the initial values.