Feat: Completed functionality of CatClicker Premium Pro
Finished up the functionality - mainly the display of tha admin
section and the save and cancel buttons within.

Also fixed an issue that cropped up where submitting the admin form
would overwrite the integer in the clickout with a string, and then
additional clicks would just concatenate a 1 onto that string rather
then doing an integer add, so had to convert the string back to a number
in the octopus.submitForm function.